separate the "local" commands
